Move enums to src/enums.ts
