--- src/BulletChartOrientation.ts
@@ -1,6 +1,0 @@
-export enum BulletChartOrientation {
-    HorizontalLeft = "HorizontalLeft",
-    HorizontalRight = "HorizontalRight",
-    VerticalTop = "VerticalTop",
-    VerticalBottom = "VerticalBottom"
-}
