+++ src/enums.ts
@@ -1,0 +1,15 @@
+export enum BarRectType {
+    Minimum = "Minimum",
+    NeedsImprovement = "NeedsImprovement",
+    Satisfactory = "Satisfactory",
+    Good = "Good",
+    VeryGood = "VeryGood",
+    Bullet = "Bullet"
+}
+
+export enum BulletChartOrientation {
+    HorizontalLeft = "HorizontalLeft",
+    HorizontalRight = "HorizontalRight",
+    VerticalTop = "VerticalTop",
+    VerticalBottom = "VerticalBottom"
+}
